refactor(gui): 阶段 B 目录重组 —— tree/ 提升到顶层
把 panel/tree/、widget/layout/、widget/node.rs、widget/desc.rs 迁移
到新的 gui/src/tree/ 顶层模块,并把 PanelTree 重命名为 Tree,让"树"
作为独立基础设施脱离 panel,为阶段 C(canvas 分支共享同一棵树、消除
PanelLayer)做技术前置。

文件迁移映射:
- widget/layout/ 整目录 → tree/layout/
- widget/node.rs → tree/node.rs(PanelNode 结构名不变)
- widget/desc.rs → tree/desc.rs
- panel/tree/tree.rs → tree/tree.rs(PanelTree → Tree 重命名)
- panel/tree/layout.rs → tree/layout_adapter.rs(拆分:layout wrapper
  删除、scroll 独立为 tree/scroll.rs、LayoutTree impl 留在
  layout_adapter.rs)
- panel/tree/diff.rs → tree/diff.rs
- panel/tree/paint.rs → tree/paint.rs
- panel/tree/hit.rs → tree/hit.rs
- panel/tree/mod.rs 内容合并到 tree/mod.rs 后删除

新建文件:
- tree/mod.rs(模块入口,薄风格 re-export)
- tree/scroll.rs(impl Tree { scroll } method 形式)

删除:
- canvas/tree/ 整目录(6 个空注释文件,阶段 C 设计已改为"一棵树两个
  分支",tree 在顶层)
- panel/tree/ 整目录(所有文件搬走后)
- widget/mod.rs 里的 pub use layout::{...}(死代码,无引用)
- panel/tree/layout.rs 里的 pub fn layout wrapper(纯冗余代理)

API 变化:
- PanelTree → Tree(结构体重命名,所有 call site 同步)
- scroll 从 free function 变成 impl Tree method:
  scroll(&mut tree, id, delta) → tree.scroll(id, delta)
- pub mod layout + pub use layout::layout 共存,外部可用
  crate::tree::layout::BoxStyle(mod)和 crate::tree::layout(&mut tree)(fn)

已知 TODO(不在阶段 B 范围):
- tree/node.rs 和 tree/desc.rs 仍依赖 widget::props::WidgetProps,
  形成 tree → widget 的反向依赖,违反目标文档 §5 的"tree 不依赖
  widget"原则。阶段 C 将通过架构调整解决。

零行为变化:demo 视觉和交互与阶段 A 结束时(5c08791)完全一致。
所有变更纯为文件位置、模块路径和类型名称调整。

Refs:
- docs/target/2.0.0-gui.md §5 目标文件结构
- docs/superpowers/specs/2026-04-11-phase-b-tree-reorganization-design.md
- docs/superpowers/plans/2026-04-11-phase-b-tree-reorganization.md
