--- gui/src/canvas/tree/diff.rs
@@ -1,1 +1,0 @@
-// 引擎数据同步
