--- gui/src/canvas/tree/hit.rs
@@ -1,1 +1,0 @@
-// 画布命中检测
