--- gui/src/canvas/tree/mod.rs
@@ -1,1 +1,0 @@
-// 画布元素树导出
