--- gui/src/canvas/tree/node.rs
@@ -1,1 +1,0 @@
-// NodeView / ConnectionView 渲染状态
