--- gui/src/canvas/tree/paint.rs
@@ -1,1 +1,0 @@
-// 画布绘制遍历
