--- gui/src/canvas/tree/tree.rs
@@ -1,1 +1,0 @@
-// 画布元素存储（节点卡片、连线）
