+++ gui/src/panel/renderer.rs
@@ -1,14 +1,14 @@
-use super::tree::{reconcile, layout, paint, hit_test, scroll, Desc, PanelTree, NodeId};
+use crate::tree::{reconcile, layout, paint, hit_test, Desc, NodeId, Tree};
 use crate::renderer::{Rect, Renderer, TextMeasurer};
 
 /// 面板渲染器。封装 reconcile → layout → paint 流水线。
 pub struct PanelRenderer {
-    tree: PanelTree,
+    tree: Tree,
 }
 
 impl PanelRenderer {
     pub fn new() -> Self {
-        Self { tree: PanelTree::new() }
+        Self { tree: Tree::new() }
     }
 
     /// 更新面板内容并重新布局。
@@ -39,14 +39,14 @@ impl PanelRenderer {
 
     /// 滚动指定节点。
     pub fn scroll(&mut self, node_id: NodeId, delta: f32) {
-        scroll(&mut self.tree, node_id, delta);
+        self.tree.scroll(node_id, delta);
     }
 
     pub fn root(&self) -> Option<NodeId> {
         self.tree.root()
     }
 
-    pub fn tree(&self) -> &PanelTree {
+    pub fn tree(&self) -> &Tree {
         &self.tree
     }
 }
