--- gui/src/panel/tree/mod.rs
@@ -1,13 +1,0 @@
-mod diff;
-mod hit;
-mod layout;
-mod paint;
-mod tree;
-
-pub use crate::widget::desc::Desc;
-pub use crate::widget::node::NodeId;
-pub use diff::reconcile;
-pub use hit::hit_test;
-pub use layout::{layout, scroll};
-pub use paint::paint;
-pub use tree::PanelTree;
