+++ gui/src/tree/hit.rs
@@ -1,11 +1,11 @@
-use crate::widget::node::NodeId;
-use super::tree::PanelTree;
+use super::node::NodeId;
+use super::tree::Tree;
 
-pub fn hit_test<'a>(tree: &'a PanelTree, root: NodeId, x: f32, y: f32) -> Option<&'a str> {
+pub fn hit_test<'a>(tree: &'a Tree, root: NodeId, x: f32, y: f32) -> Option<&'a str> {
     hit_test_node(tree, root, x, y)
 }
 
-fn hit_test_node<'a>(tree: &'a PanelTree, node_id: NodeId, x: f32, y: f32) -> Option<&'a str> {
+fn hit_test_node<'a>(tree: &'a Tree, node_id: NodeId, x: f32, y: f32) -> Option<&'a str> {
     let Some(node) = tree.get(node_id) else { return None };
 
     let r = &node.rect;
