+++ gui/src/tree/mod.rs
@@ -1,0 +1,17 @@
+mod desc;
+mod diff;
+mod hit;
+pub mod layout;
+mod layout_adapter;
+mod node;
+mod paint;
+mod scroll;
+mod tree;
+
+pub use desc::Desc;
+pub use diff::reconcile;
+pub use hit::hit_test;
+pub use layout::layout;
+pub use node::{NodeId, NodeKind, PanelNode};
+pub use paint::paint;
+pub use tree::Tree;
