+++ gui/src/widget/atoms/dropdown.rs
@@ -22,8 +22,8 @@ impl WidgetProps for DropdownProps {
         fmt::Debug::fmt(self, f)
     }
     fn build(&self, id: &str) -> WidgetBuild {
-        use crate::widget::desc::Desc;
-        use crate::widget::layout::{BoxStyle, Decoration, Size, Direction, Align, Edges, LeafKind};
+        use crate::tree::Desc;
+        use crate::tree::layout::{BoxStyle, Decoration, Size, Direction, Align, Edges, LeafKind};
         use crate::renderer::{Border, Color};
 
         // shadcn zinc 色系
