+++ gui/src/widget/atoms/text_input.rs
@@ -21,8 +21,8 @@ impl WidgetProps for TextInputProps {
         fmt::Debug::fmt(self, f)
     }
     fn build(&self, id: &str) -> WidgetBuild {
-        use crate::widget::desc::Desc;
-        use crate::widget::layout::{BoxStyle, Decoration, Size, Direction, Edges, LeafKind};
+        use crate::tree::Desc;
+        use crate::tree::layout::{BoxStyle, Decoration, Size, Direction, Edges, LeafKind};
         use crate::renderer::{Border, Color};
 
         // shadcn zinc 色系
