+++ gui/src/widget/atoms/toggle.rs
@@ -21,8 +21,8 @@ impl WidgetProps for ToggleProps {
         fmt::Debug::fmt(self, f)
     }
     fn build(&self, id: &str) -> WidgetBuild {
-        use crate::widget::desc::Desc;
-        use crate::widget::layout::{BoxStyle, Decoration, Size, Direction, Align, Justify, Edges, LeafKind};
+        use crate::tree::Desc;
+        use crate::tree::layout::{BoxStyle, Decoration, Size, Direction, Align, Justify, Edges, LeafKind};
         use crate::renderer::Color;
 
         // shadcn 色系
