+++ gui/src/widget/mod.rs
@@ -1,7 +1,4 @@
 pub mod atoms;
-pub mod desc;
-pub mod layout;
-pub mod node;
 pub mod props;
 
 pub mod action;
@@ -9,5 +6,3 @@ mod focus;
 mod mapping;
 pub mod state;
 mod text_edit;
-
-pub use layout::{BoxStyle, Decoration, LeafKind, LayoutTree, Size};
